Use npm exec instead of node to run tsc
The files that pnpm installs in node_modules/.bin are not js files, so running
`node node_modules/.bin/tsc` doesn't actually work. Instead, this runs `npm
exec tsc` which should work in all environments.

This also gets rid of the `max-old-space-size` options, but those are being
handled in the `env` variables of the calling github action now.